Refactor Trilinos' linear op. stuff into own header
