+++ include/deal.II/lac/linear_operator.h
@@ -1237,110 +1237,6 @@ linear_operator(const OperatorExemplar &operator_exemplar, const Matrix &matrix)
 //@}
 
 
-#ifdef DEAL_II_WITH_TRILINOS
-
-// Forward declarations:
-
-namespace TrilinosWrappers
-{
-  class SparseMatrix;
-  class PreconditionBase;
-
-  namespace internal
-  {
-    namespace LinearOperator
-    {
-      class TrilinosPayload;
-    }
-  }
-}
-
-namespace TrilinosWrappers
-{
-  /**
-   * @name Creation of a LinearOperator
-   */
-//@{
-
-
-  /**
-   * @relates LinearOperator
-   *
-   * Return a LinearOperator that is the identity of the vector space @p Range.
-   *
-   * This function is the equivalent of the dealii::identity_operator, but
-   * ensures full compatibility with Trilinos operations by preselecting the
-   * appropriate template parameters.
-   *
-   * @author Jean-Paul Pelteret, 2016
-   *
-   * @ingroup TrilinosWrappers
-   */
-  template <typename Range>
-  inline LinearOperator<Range, Range, TrilinosWrappers::internal::LinearOperator::TrilinosPayload>
-  identity_operator(const std::function<void(Range &, bool)> &reinit_vector)
-  {
-    typedef TrilinosWrappers::internal::LinearOperator::TrilinosPayload Payload;
-    return dealii::identity_operator<Range, Payload>(reinit_vector);
-  }
-
-
-  /**
-   * @relates LinearOperator
-   *
-   * A function that encapsulates generic @p matrix objects, based on an
-   * @p operator_exemplar, that act on a compatible Vector type into a
-   * LinearOperator.
-   *
-   * This function is the equivalent of the dealii::linear_operator, but
-   * ensures full compatibility with Trilinos operations by preselecting the
-   * appropriate template parameters.
-   *
-   * @author Jean-Paul Pelteret, 2016
-   *
-   * @ingroup TrilinosWrappers
-   */
-  template <typename Range, typename Domain = Range,
-            typename Matrix>
-  inline LinearOperator<Range, Domain, TrilinosWrappers::internal::LinearOperator::TrilinosPayload>
-  linear_operator(const TrilinosWrappers::SparseMatrix &operator_exemplar, const Matrix &matrix)
-  {
-    typedef TrilinosWrappers::SparseMatrix OperatorExemplar;
-    typedef TrilinosWrappers::internal::LinearOperator::TrilinosPayload Payload;
-    return dealii::linear_operator<Range, Domain, Payload, OperatorExemplar, Matrix>(operator_exemplar, matrix);
-  }
-
-
-  /**
-   * @relates LinearOperator
-   *
-   * A function that encapsulates generic @p matrix objects that act on a
-   * compatible Vector type into a LinearOperator.
-   *
-   * This function is the equivalent of the dealii::linear_operator, but
-   * ensures full compatibility with Trilinos operations by preselecting the
-   * appropriate template parameters.
-   *
-   * @author Jean-Paul Pelteret, 2016
-   *
-   * @ingroup TrilinosWrappers
-   */
-  template <typename Range, typename Domain = Range>
-  inline LinearOperator<Range, Domain, TrilinosWrappers::internal::LinearOperator::TrilinosPayload>
-  linear_operator(const TrilinosWrappers::SparseMatrix &matrix)
-  {
-    typedef TrilinosWrappers::SparseMatrix Matrix;
-    typedef TrilinosWrappers::internal::LinearOperator::TrilinosPayload Payload;
-    return dealii::linear_operator<Range, Domain, Payload, Matrix, Matrix>(matrix, matrix);
-  }
-
-//@}
-
-} // namespace TrilinosWrappers
-
-#endif
-
-
 DEAL_II_NAMESPACE_CLOSE
 
 #endif // DEAL_II_WITH_CXX11
