+++ include/deal.II/lac/linear_operator_tools.h
@@ -16,12 +16,16 @@
 #ifndef dealii__linear_operator_tools_h
 #define dealii__linear_operator_tools_h
 
-// many usage cases lead to a combination of LinearOperator and
-// PackagedOperation. To ease the pain of reading compilation errors, just include
-// all headers we ever need to use LO and friends in one place:
+// Many usage cases lead to a combination of LinearOperator and
+// PackagedOperation. To ease the pain of reading compilation errors, just
+// include all headers we ever need to use LO and friends in one place:
+
 #include <deal.II/lac/linear_operator.h>
 #include <deal.II/lac/block_linear_operator.h>
+#include <deal.II/lac/trilinos_linear_operator.h>
+
 #include <deal.II/lac/packaged_operation.h>
+
 #include <deal.II/lac/constrained_linear_operator.h>
 #include <deal.II/lac/schur_complement.h>
